This was initially posted to Google cloud.
(2018-11-27 03:50) Initial commit of the Themis RC 1.0 on Google Cloud.
This will build with Version: 8.2.0.00007 of the CCS (Code Composer Studio) tools from TI.
This build includes the 1.06 of c2000ware as it pertains to this project plus version 3.1 of the DCL (Digital Control Library) from TI.
I tested this build on a Themis board before committing - and it runs - USB prints work and the NMEATime Panel is getting a good lock.

To Do:  Source code - mainly the themis_main.c file needs to be cleaned up somewhat.
Note this source also contains a grass roots PID in lieu of what's in the TI DCL that runs pretty good.
The DCL is included for comparison/testing/enhancements.
--- jcw
(2018-12-13 01:01)Built with: Version: 8.3.0.00009; recently released by TI.
Note - to use the USB peripherals on the DSP and uBlox - USB drivers may need to be downloaded from those companies.
--- jcw